make select operand component
